test(bootstrap): move derive test
